+++ ReduxCore/framework.php
@@ -124,6 +124,7 @@
                         self::$_url     = trailingslashit( self::$wp_content_url . $relative_url );
                     }
                 }
+
             }
 
             // ::init()